+++ CHANGELOG.md
@@ -27,6 +27,15 @@ All notable changes to this project will be documented in this file.
 
 - [BREAKING] `--algorithm` option is removed from `iexec dataset encrypt`
 - [BREAKING] removed `iexec.network.id` and `iexec.network.isSidechain`, use `iexec.network.getNetwork() => Promise<{chainId: String, isSidechain: Boolean}>`
+- [BREAKING] drop previously deprecated `iexec wallet show --raw` returned json key `balance.ETH`, use `balance.ether` instead
+- [BREAKING] drop previously deprecated `bridge.bridgedChainId` in `chain.json` use `bridge.bridgedChainName` instead
+- [BREAKING] drop previously deprecated `iexec.orderbook.fetchWorkerpoolOrderbook(category, options)` use `category` as an option of `iexec.orderbook.fetchWorkerpoolOrderbook(options)`
+- [BREAKING] drop previously deprecated `iexec.orderbook.fetchRequestOrderbook(category, options)` use `category` as an option of `iexec.orderbook.fetchRequestOrderbook(options)`
+- [BREAKING] drop previously deprecated `iexec.orderbook.fetchAppOrderbook()` returned value `appOrders` use `orders`
+- [BREAKING] drop previously deprecated `iexec.orderbook.fetchDatasetOrderbook()` returned value `datasetOrders` use `orders`
+- [BREAKING] drop previously deprecated `iexec.orderbook.fetchWorkerpoolOrderbook()` returned value `workerpoolOrders` use `orders`
+- [BREAKING] drop previously deprecated `iexec.orderbook.fetchRequestOrderbook()` returned value `requestOrders` use `orders`
+- [BREAKING] drop previously deprecated `task.waitForTaskStatusChange(taskid, initialStatus)` use `task.obsTask(taskid)`
 
 ## [5.3.0] 2021-05-05